♻️ Extract Selectors
